Add BluetoothDevice::VendorIDSource()
The Vendor ID of a device can be either allocated by the Bluetooth SIG
or USB IF, providing two completely seperate overlapping namespaces for
identifiers.

In order to distinguish which vendor an identifier represents, add a
VendorIDSource() method to BluetoothDevice that returns BLUETOOTH or
USB respectively.

Add to the metrics collection as well.

BUG=350432
TEST=device_unittests
R=armansito@chromium.org, isherman@chromium.org

Review URL: https://codereview.chromium.org/191223003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@256344 4ff67af0-8c30-449e-8e8b-ad334ec8d88c