Re-land r252832 after revert r252852 addressing Andriod test failure.
The bookmark undo is not desired functionality on Android and so the undo framework code is not compiled for Android.

This means that the group undo events are not fired on Android and thus the test should not call ExpectGroupedChangeCount() on Android.

Originally Committed: r252832
Reverted: r252852 ( Review URL: https://codereview.chromium.org/176563002 )

BUG=126092

Review URL: https://codereview.chromium.org/191943002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@256506 4ff67af0-8c30-449e-8e8b-ad334ec8d88c