[Fullscreen-Within-Tab] Fix drag-out of captured fullscreen tabs.
This change allows FullscreenControllers in separate Browsers to hand-off the FWT state for tabs being dragged between browser windows.  This is achieved by utilizing the base::SupportsUserData mechanism to associate the FWT state with the WebContents instance for the tab.

Mac: Minor change to the TabContentsController inititializer to correct for the hand-off of FWT's view between TabContentsControllers in separate browser windows.

BUG=348423

Review URL: https://codereview.chromium.org/190763006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@256515 4ff67af0-8c30-449e-8e8b-ad334ec8d88c