Properly disable background mode when deleting profiles.
Updated BackgroundModeManager to recalculate the background state when a
profile is removed - this avoids crashes where we are in background mode but
there are no background profiles.

Changed BackgroundApplicationListModel to complain if a caller forgets to
remove the observer before the model is destroyed. This would have caught
this bug much sooner had it been enabled.

BUG=346214

Review URL: https://codereview.chromium.org/192943002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@256539 4ff67af0-8c30-449e-8e8b-ad334ec8d88c