mac: History swiping doesn't work right with iframes.
In the previous behavior, it was possible for a user to start scrolling an
iframe with the touch pad, and then trigger history navigation. This happened
because the Cocoa view only received callbacks for whether an event in a
gesture wasn't handled. If any event in a gesture happened to not be handled by
blink, then that event could trigger history swiping, even though blink may
have handled earlier events in the same gesture!

My solution is to propagate the information for whether an event was handled to
the Cocoa view. History swiping is only enabled if no event is ever handled by
blink.

BUG=321437
TEST=Navigate to www.google.com. Then navigate to
http://codepen.io/anon/pen/ujsec. Place the cursor in the box that says "place
cursor here", and start a downwards 2-finger swipe. Half way through the swipe,
change directions and swipe to the left (or right, depending on your settings)
to navigate backwards in history. Prior to this change, the history swipe arrow
would appear. In this change, the history swipe arrow no longer appears.

Review URL: https://codereview.chromium.org/194713016

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@256929 4ff67af0-8c30-449e-8e8b-ad334ec8d88c