Removed redundant notifications info from about:sync-internals
This patch removes the tab 'Notifications' from about:sync-internals.

This tab used to be useful for debugging notifications as they arrived, 
but now that its functionality has been replaced and augmented 
in the new about:invalidations, with more general functionality and extended 
logging information, it is no longer necessary to have it here.

BUG=351559

Review URL: https://codereview.chromium.org/196373004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@257182 4ff67af0-8c30-449e-8e8b-ad334ec8d88c