Update frame color API to use unions.
This allows us to use the 'frame' option either as a string (for type)
or as a dictionary (for type and / or color).

Note, this is backwards incompatible with the previous way of setting frame color. To use frame color you now use 'frame' instead of 'frameOptions'.

BUG=339558

Review URL: https://codereview.chromium.org/197813008

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@257390 4ff67af0-8c30-449e-8e8b-ad334ec8d88c