[Mac] Manage extension popups more reasonably.
There were two opportunities for an ExtensionPopupController to be released,
in +[ExtensionPopupController showURL:...], which called -close and
-autorelease, and in -[BaseBubbleController windowWillClose:], which would be
triggered by -close and would also call -autorelease. This result in
over-freeing the object, which was observed causing ZombieObjectCrashes
(typically -windowDidResignKey: was the first message sent to the deallocated
ExtensionPopupController).

The main active ingredient in this change is removing the extra -autorelease
in +[ExtensionPopupController showURL:...]. Additional care is taken regarding
management of extension_popup_controller.mm's gObject. Some cleanup tasks in
BaseBubbleController are moved from -close to -windowWillClose:, because
-close was observed as being too early.

The popup identity comparison in ExtensionViewHostMac was simplified to only
consider an ExtensionPopupController as belonging to an EVHM if the object
matched, without comparing the extension ID. The extension ID check could make
it close an unrelated popup when repeatedly clicking on an extension icon.

TEST=Repeatedly click on an extension icon. Vary the speed, about 1Hz and
     faster. Doing this for 30 seconds would be enough to see a
     ZombieObjectCrash or other bad behavior (such as a
     disappearing-reappearing main browser window).
BUG=351278
R=rsesek@chromium.org

Review URL: https://codereview.chromium.org/200513005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@257641 4ff67af0-8c30-449e-8e8b-ad334ec8d88c