Add 'View in Web Store' link to the App Info dialog
Add a link that says 'View in Web Store' to the App Info dialog. When clicked,
it opens the link to the app in the web store (but does not close the dialog).

Does not display the link for apps that are not from the web store.

BUG=266739

Review URL: https://codereview.chromium.org/193633003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@257733 4ff67af0-8c30-449e-8e8b-ad334ec8d88c