Add --disable-low-res-tiling. Disable low res tiling to save power.
Generating low resolution tiles uses about 10% of the power used by
pages that have gif animations (measured on Nexus 4). These are the
most common pages on which chrome consumes above baseline power.

It is unclear whether low resolutions are still valuable. It is
likely that having low resolution tiles is more useful on older
devices. For the moment providing this as a flag, so that it
is easier to investigate whether dropping low resolution tiles
is reasonable. Default behavior is unchanged, low resolution tiles
are enabled.

BUG=

Review URL: https://codereview.chromium.org/196473007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@257790 4ff67af0-8c30-449e-8e8b-ad334ec8d88c