remove obsolete webview code from signin header helper
and clear the isolated partition used by chrome sign in page upon the signin extension is unloaded, because Gaia requires Chrome signin to be executed with a clean cookie jar. See crbug/349461 for an example when sign in is broken with a dirty cookie jar.

BUG=349461

Review URL: https://codereview.chromium.org/187923002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@257805 4ff67af0-8c30-449e-8e8b-ad334ec8d88c