Add the notification settings to the access request data.
When generating an access request, we copy the notification
setting of the custodian (which is a shared setting) to the
access request data. This allows to use this setting to
decide in the Chrome Sync server if this access request
should trigger a notification or not.

BUG=353543
TEST=Create an access request from a supervised user profile. Nothing should break.

Review URL: https://codereview.chromium.org/202623005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@257902 4ff67af0-8c30-449e-8e8b-ad334ec8d88c