Disable CoreAnimation by default
We're seeing lots of OOM crashes in dev and canary. See if this makes
them go away.

Make sure the TabStripView always gets a layer (regardless of the
flag), since we'll need the layer to use the CoreAnimation tab spinners,
which will be staying in.

BUG=245900

Review URL: https://codereview.chromium.org/203573005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@257924 4ff67af0-8c30-449e-8e8b-ad334ec8d88c