Allow Goodies page to access echo.checkModelEligibility, and all extensions to access echo.checkEligibility.
Since this external_connectable feature does not allow to specify permission per
method level. Any caller whitelisted here are allowed to call echo. So in
addition to this whitelist, the actual echo extension code also checks if the
caller is legitimate. As a result of current config, only our goodies page can
get result from echo.checkModelEligibility.

ids:[*] was the default value if we don't configure this externally_connectable.
so adding it will not change the existing behavior.

On the other hand, checkEligibility is already accessible by all webpages and
extensions, it's up to the server side to reject the request if the origin is
not legitimate.

BUG=chromium:352446
TEST=browser test

Review URL: https://codereview.chromium.org/196283022

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@258072 4ff67af0-8c30-449e-8e8b-ad334ec8d88c