[Cocoa] TabContentsController will size embedded fullscreen views differently during tab capture.
This is the same change as http://codereview.chromium.org/203093002/, but for the Mac UI: During tab capture sessions, fullscreened views will be displayed within the tab contents area at the capture video resolution.  This prevents unwanted scaling and letterboxing, for a better user experience on the receiving screen (e.g., a Chromecast).

Also made clean-ups to many of the comments in the TabContentsController header file to clarify how it is used and how it behaves internally.

BUG=256870

Review URL: https://codereview.chromium.org/205473002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@258244 4ff67af0-8c30-449e-8e8b-ad334ec8d88c