Use EXTENSION_REMOVED to reset IconImage's extension pointer
EXTENSION_REMOVED is more suitable than EXTENSION_UNLOADED, as
the purpose of resetting the extension_ pointer is to ensure it's not
used after the extension gets removed. EXTENSION_UNLOADED does not
match that task, as it may not be dispatched when an extension gets
removed (e.g. for disabled extension).
Also, EXTENSION_UNLOADED may be dispatched in few situation when it's
not desirable to invalidate IconImage's image loading (e.g. when an
extension gets disabled)

Also, reload app list icon when disabled extension gets updated (as in
this case there is no EXTENSION_LOADED notification)

BUG=343978,277907

Review URL: https://codereview.chromium.org/195793025

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@258443 4ff67af0-8c30-449e-8e8b-ad334ec8d88c