[Mac, Win] Show a user manager tutorial once per profile.
This tutorial is only shown when the user explicitly brings up the
user manager, and not, for example, as a result of locking the profile.
This preference can be made syncable in the future.

WebUI change: if there is only one pod, show the tutorial to the
right side of the pod, rather than centered (which is how it is displayed
for multiple pods).

BUG=324052
TEST=Start Chrome with --new-profile-management enabled. Press "View
all users". You should see the tutorial. Close the user manager and
repeat the previous step. You shouldn't see the tutorial again.

Review URL: https://codereview.chromium.org/199533004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@258483 4ff67af0-8c30-449e-8e8b-ad334ec8d88c