Prompts the user whether or not to retry loading an unpacked extension if the extension failed to load.
Sends a retry notification from extension service to the settings handler on load error. The notification contains the user's response to the retry prompt.

Modified ReportError to prompt a Question box instead of a Warning box when a |user_response| is expected.

Maintains a list of all currently loading extensions. The retry notification listener is only active while extensions are in the process of loading (when the list is not empty).

BUG=338899

Review URL: https://codereview.chromium.org/167323003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@258677 4ff67af0-8c30-449e-8e8b-ad334ec8d88c