Stop PasswordStoreMac thread in Shutdown().
Currently PasswordStoreMac creates a supplementary thread in Init(), which is called on the UI thread. PasswordStoreMac stops the thread in its destructor. That's a problem if the destructor is not called on the UI thread. The correct way is to stop the thread in Shutdown(), as that's always called on the UI thread.

BUG=350872

Review URL: https://codereview.chromium.org/206503003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@258927 4ff67af0-8c30-449e-8e8b-ad334ec8d88c