net: more pinning debugging.
The reports from the last round of debugging suggests that the pins are being
calculated correctly and that the Google pins are correct in the binary. Since
the reports mostly have "apis.google.com" as the hostname, suspicion falls on
malware that might be changing the pinning entry for just that hostname.

BUG=none

Review URL: https://codereview.chromium.org/210073003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@259078 4ff67af0-8c30-449e-8e8b-ad334ec8d88c