Adjust enrollment screen strings for forced re-enrollment.
This is major overhaul of the enrollment script HTML/JS in favor of
a more declarative way of doing things as opposed to the UI element
fiddling the old code performed. Major visual change is adding a right
margin pane to the enrollment sign-in screen similar to the login
screen.

BUG=chromium:356143
TEST=Enrollment screen still works for FRE, AE, and manal enrollment.
R=joaodasilva@chromium.org, ygorshenin@chromium.org

Review URL: https://codereview.chromium.org/212113003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@259597 4ff67af0-8c30-449e-8e8b-ad334ec8d88c