Do not launch the feedback app from an OTR profile.
chrome::ShowFeedbackPage was dispatching an event to the feedback API
on the browser's current profile (or the global last used profile)
directly.

This can be an OTR profile and lead to bad behavior. App windows are
never hosted on OTR profiles, so the feedback app window will not keep
the OTR profile alive. The feedback app can therefore outlive the OTR profile
which launched it, resulting in operations on browser context services
(like FeedbackPrivateAPI, ExtensionSystem, ProcessManager, etc) which no
longer exist.

This change causes ShowFeedbackPage to resolve an OTR profile to its
original profile before dispatching an event to the feedback app.

BUG=354546

Review URL: https://codereview.chromium.org/210353002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@259781 4ff67af0-8c30-449e-8e8b-ad334ec8d88c