[fsp] Simplify ID of provided file systems.
Previously, the ID was very lenghty, being same as the mount point name. Also,
it contained the extension id, which was too much.

To makes things simple, the ID has been changed to an int only, which is
consistent with setTimeout/setInterval as well as upcoming request's id,
which are just positive numbers.

TEST=unit_tests, browser_tests
BUG=248427

Review URL: https://codereview.chromium.org/209663006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@259842 4ff67af0-8c30-449e-8e8b-ad334ec8d88c