Support Ethernet in WifiConfigView
This is a short term solution that allows us to use the existing WifiConfigView dialog to configure ethernet. Rather than renaming the class and/or introducing a helper class, I decided to keep the changes to a minimum since we are planning to replace this class entirely soon.

BUG=344011
R=pneubeck@chromium.org, tbarzic@chromium.org

Review URL: https://codereview.chromium.org/208813005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@259870 4ff67af0-8c30-449e-8e8b-ad334ec8d88c