Fix build when ENABLE_MANAGED_USERS isn't defined
In https://codereview.chromium.org/137753012/, (and its associated reverts)
ManagedUserSigninManagerWrapper is added to create a bridge between managed
users and normally authenticated users.  Unfortunately, in builds where
enable_managed_users==0 in gyp, the wrapper is excluded from the build due
to an exclusion list in chrome_browser.gypi.

Therefore, when enable_managed_users==0, libchromeshell fails to build
(as well as any other projects that pull in chrome.gyp:chrome).

1) This moves ManagedUserSigninManagerWrapper to chrome/browser/sync because
it is the only place this wrapper is used.  This also prevents the class from
being conditionally excluded due to the rules in chrome_browser.gypi.

2) This commit also adds a missing #ifdef in tab_helpers.cc, which resulted
in a compilation error (unused variable 'profile') if ENABLE_MANAGED_USERS is
not defined.

BUG=352733

Review URL: https://codereview.chromium.org/200573003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@259928 4ff67af0-8c30-449e-8e8b-ad334ec8d88c