CoreAnimation: Give app windows layers
On 10.6-10.8, there is a brief flicker of corruption when the
RWHVCocoa (which has a CALayer) is added to the already-visible
window's NSView hierarchy. I suspect that this is a mini-seizure
that the window system has when transitioning the window from
a "does not have layers" mode to a "has layers" mode. Adding a
layer to the window's NSView hierarchy before the window is made
visible makes this seizure not happen.

BUG=356627

Review URL: https://codereview.chromium.org/215593002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@260268 4ff67af0-8c30-449e-8e8b-ad334ec8d88c