Resetting metrics ids on clump detection.
When a cloned install is detected the metrics client id and low entropy source are reset on the next start of Chrome. This is done by setting a pref in local state. The next time either the client id or low entropy source are loaded from the local state, they are first reset.

BUG=343273
NOTRY=true

Review URL: https://codereview.chromium.org/213363004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@260348 4ff67af0-8c30-449e-8e8b-ad334ec8d88c