Turn on and use the AppInfo data.
Synced Notifications uses the SyncedNotificationAppInfo struct to name
the sending services and supply icons for them.  The first two checkins
in this larger change established the datatype and the sync change
processor for it.  This checkin wires up the data type to the rest of the
Synced Notifications code.

This checkin now uses the SNAI class to provide the bitmap and service name,
and removes the hardcoding to our first service name provider and bitmap.
This checkin also adds code to fetch the bitmaps provided by the SNAI
protobuf.

New unit tests have been added to cover the new functionality.

A lifetime dependency has been set up to ensure that the
SyncedNotificationAppInfoService is always present whenever we have a
ChromeNotifierService.

BUG=280266

Review URL: https://codereview.chromium.org/193773003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@260433 4ff67af0-8c30-449e-8e8b-ad334ec8d88c