Prevent SHChangeNotify on app list drag on Windows.
This CL fixes an issue with all taskbar icons flashing on every app list
drag due to a SHChangeNotify being fired. This was being caused an overwrite
of the web app path shortcut which is created for drag and drop to the taskbar.
The fix skips the shortcut creation if it already exists (but still updates the icon).

This CL also extracts a GetIconFilePath function.

BUG=356961

Review URL: https://codereview.chromium.org/213833005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@260439 4ff67af0-8c30-449e-8e8b-ad334ec8d88c