Measure IOThreadCreation to request start
Adds UMA to measure the lag from the time the IOThread
is created to the time the first request is started. This will be helpful to measure the benefit of starting a request in the browser process before the render process initiates the request.

BUG=354438

Review URL: https://codereview.chromium.org/214123002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@260593 4ff67af0-8c30-449e-8e8b-ad334ec8d88c