sync: Prepare for integration test refactoring
Adds an optional virtual method to sync tests to control whether or not
self-notifications should be enabled for a specific test.  Defaults to
on.

Introduces the multi-client status change checker.  This can be used to
wait for a specific condition while waiting listening to change events
emitted by several different ProfileSyncServices.

Adds some helper functions to SyncTest that can be used to fetch the
active ProfileSyncServices.  This functions are helpful for initializing
StatusChangeCheckers.

These changes should have no effect on existing tests.  They will be
used to implement a new style of integration test that does not rely on
sync-cycle concepts like "quiescence" or the self-notifications that
those concepts depend on.

BUG=97780,95742

Review URL: https://codereview.chromium.org/215583002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@260616 4ff67af0-8c30-449e-8e8b-ad334ec8d88c