[Files.app] Use getDriveEntryProperties() to retrieve metadata
Previously, we used FileEntry.getMetadata() to retrieve filesystem metadata and getDriveEntryProperties() to get Drive metadata, so we need 2 calls for 1 file on Drive.

With this patch, getDriveEntryProperties() returns not only Drive metadata but also filesystem metadata. It's enough to call only getDriveEntryProperties() and we can reduce a number of calls by half.

BUG=345196
TEST=browser_test passes.
R=asargent@chromium.org, hashimoto@chromium.org, hirono@chromium.org

Review URL: https://codereview.chromium.org/195763004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@260688 4ff67af0-8c30-449e-8e8b-ad334ec8d88c