jankometer: Use the stub implementation for non-Windows.
The jankometer implementation only does anything on Windows. So use
the stub implementation (used on mac, android) for linux-aura. Also,
remove the gtk related code from jankometer.cc, and rename it to
jankometer_win.cc to make it clear that it is used only on Windows.

BUG=358751
R=sky@chromium.org

Review URL: https://codereview.chromium.org/213673007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@261189 4ff67af0-8c30-449e-8e8b-ad334ec8d88c