Fix bitmap fetcher race condition.
There is no need to re-fetch a bitmap if we are already fetching it.
The code can get into this race condition when processing synced notifications,
with one notification being loaded from sync stored data, and a newer version
of the same notification coming in from the wire (from the sync server).

BUG=358240

Review URL: https://codereview.chromium.org/218633005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@261213 4ff67af0-8c30-449e-8e8b-ad334ec8d88c