Add non-blocking sync code to ProfileSyncService
Adds support for non-blocking sync to the ProfileSyncService and related
classes.  This continues the work of r258390 and r259921.  Like those
patches, it is not expected to have any impact on behavior.

Introduces ProfileSyncService::RegisterNonBlockingType().  This function
will act as an alternative to RegisterDataTypeController().

Adds some support for these non-blocking types.  Adds a special
case to the code that generates the 'type status' table on the
about:sync page.

Instantiates and copies a SyncCoreProxy object to the ProfileSyncService
as backend initialization completes.  This will be an important part of
non-blocking data type initialization, once that functionality is
supported.

BUG=351005

Review URL: https://codereview.chromium.org/217183003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@261590 4ff67af0-8c30-449e-8e8b-ad334ec8d88c