Add support for hooking up external feedback mechanism for DOM Distiller
This adds support for injecting an external feedback reporting library for
the DOM Distiller. By default, a no-op version is used.

For feedback, it is reasonable to assume that the original URL which was
distilled is needed, and this CL also adds support for retrieving that
from Java.

BUG=319881

Review URL: https://codereview.chromium.org/222883003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@261688 4ff67af0-8c30-449e-8e8b-ad334ec8d88c