PermissionBubbleManager now does not cancel requests when its WebContents is destroyed.
Closing a window while a permissions bubble was still open was crashing
on the Mac.  The call to MediaStreamDevicesController::Cancelled() calls
MediaStreamDevicesController::Deny(), which then saves the new setting.
Saving the new setting triggers an observer callback which attempts to
access and use the current web_contents() which is now NULL.

It is safer to simply not attempt to access a request once the WebContents
has been destroyed than to try to control the destruction order.

BUG=None

Review URL: https://codereview.chromium.org/224213002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@261933 4ff67af0-8c30-449e-8e8b-ad334ec8d88c