base: Do not allow MessagePumpObservers to consume events.
There is currently a single message-pump observer that can consume an event,
and it does so only when the num-lock key is pressed on Chrome OS. But it
doesn't need to consume the event, because no other dispatcher/observer cares
about this event. So remove a MessagePumpObserver's ability to consume an
event. This makes the code cleaner, and easier to understand.

BUG=354062
R=darin@chromium.org, derat@chromium.org, sky@chromium.org

Review URL: https://codereview.chromium.org/223483002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@262019 4ff67af0-8c30-449e-8e8b-ad334ec8d88c