linux-aura: Prepend 'http://' to selection-clipboard from omnibox selection.
Let the views::TextfieldController know about the clipboard that the textfield
content was copied into in OnAfterCutOrCopy(), so the controller can prepend
'http://' to the correct clipboard content (when appropriate). This fixes
pasting from the selection-clipboard on linux-aura after selecting the URL in
the omnibox.

BUG=351716
R=msw@chromium.org, sky@chromium.org

Review URL: https://codereview.chromium.org/227363003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@262196 4ff67af0-8c30-449e-8e8b-ad334ec8d88c