Move reload processing from TranslateManager to TranslateTabHelper.
This reload processing is specific to the //chrome embedder flow, wherein the
LanguageDetermined IPC (which initiates translation) is not sent from the
renderer on a reload, and thus the browser-side code must explicitly listen for
reloads to initiate translation as appropriate. If the relevant renderer code
ever moves into the component, then this browser-side reload handling could
move into the content driver of the component.

BUG=359998
R=droger@chromium.org

Review URL: https://codereview.chromium.org/227363005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@262378 4ff67af0-8c30-449e-8e8b-ad334ec8d88c