Fixed quality tests after PyAuto test purge.
All test resources and reference files have been moved to be handled by
the webrtc.DEPS solution. The audio test used to find its files in the
pyauto_private folder, which has been removed. Instead I've uploaded
the files to the webrtc-resources bundle and put .sha1 files in the
webrtc test data folder, which are processed by the webrtc.DEPS
solution.

The downloading of the files is only triggered if you have the WebRTC
solution, so this should not burden people who do not want to run the
tests.

The reason for moving the files is that the tests are more healthy
if they don't have to look for resources outside the chrome/test/data
folder. I'm able to straighten out some hacks as a result.

Also fixing an issue where https://codereview.chromium.org/222873002
removed the web socket handler for the video quality test.

Depends on https://codereview.chromium.org/225263018.

BUG=360033,359579
NOTRY=true

Review URL: https://codereview.chromium.org/227633002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@262398 4ff67af0-8c30-449e-8e8b-ad334ec8d88c