[invalidations] Added a InvalidatorState last changed timestamp
This patch adds to about:invalidation a readable string next to the
InvalidatorState with the timestamp of the last time the
InvalidatorState changed. This is useful for detecting when there are
problems with the connection to the XMPP/GCM server, independantly of
Sync servers.

This patch is not a replacement of the log on the about:invalidations
page, as this timestamp is always updated even if the webpage is not
open, compared to the log that is only updated when events are relayed
in real time to the page. This new timestamp is useful for the cases
where we want to check how long has it been since we are in the same
state (like TRANSIENT_INVALIDATION_ERROR) and we could compare it to
different status timestamps in about:sync-internals or other
invalidations implementer.

BUG=263863

Review URL: https://codereview.chromium.org/222903004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@262531 4ff67af0-8c30-449e-8e8b-ad334ec8d88c