Toro: Remove chrome feature channel dep from //extensions
The existence of stable/beta/dev channel is a Chrome concept. Eliminate the reference to feature_channel.h in //extensions by observing for renderer startup and sending the channel to the renderer in //chrome.

This is also one of our last //extensions -> //chrome circular dependencies.

BUG=362244
TEST=browser_tests *Extensions*

Review URL: https://codereview.chromium.org/234183002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@263314 4ff67af0-8c30-449e-8e8b-ad334ec8d88c