Do not call GetLastUsedProfile() when gathering extension data.
We should not call GetLastUsedProfile() in UMA code that gathers the
list of installed extensions, as that GetLastUsedProfile will create a
profile if one does not yet exist.  Use a more verbose alternative that
will simply return NULL if a profile is not yet available.

This should fix an observed crash on Windows.

BUG=360467

Review URL: https://codereview.chromium.org/235563003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@263424 4ff67af0-8c30-449e-8e8b-ad334ec8d88c