Mac: Attempt to fix crasher related to tab closing animation.
It looks like -[TabStripController animationDidStopForController:finished:] was
called with a TabController that has been deallocated.
The only way I see this can happen is if -animationDidStop:finished: was called more than once.
The closing TabController is retained by _closingControllers while it's
animating the close.
The first call to will cause the TabController to be removed from
_closingControllers and deallocate, the second call will crash.
All the related crash reports are on 10.6 and 10.7, so the multiple calls could
be an OS bug that has been fixed in 10.8.

BUG=360515

Review URL: https://codereview.chromium.org/235843003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@263669 4ff67af0-8c30-449e-8e8b-ad334ec8d88c