Add language code to Wallet address.
The requestAutocomplete dialog now reads and writes the language code of
an address in Wallet. This language code is used for formatting a Wallet
address for display. The language code of a Wallet address is set based
on UI language and region-based rules when adding or editing an address.

If there's an address in Wallet that already exists, but does not have a
language code, then the address will not be updated with a language
code. Such address will be formatted with default formatting rules for a
region.

Wallet servers already support storing language code of an address.

Before this change, all Wallet addresses were formatted using the
default formatting rules for their region.

BUG=361279

Review URL: https://codereview.chromium.org/229723003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@263673 4ff67af0-8c30-449e-8e8b-ad334ec8d88c