Only run extension garbage collection when no CRX installations are currently in progress.
This should fix various race conditions where files of extensions could
get deleted while the extension was being installed.

To do this, ExtensionGarbageCollector depends on InstallTracker to be
informed whenever a CRX starts and finishes installing. To be able to
depend on InstallTracker, ExtensionGarbageCollector needed to be a
service of its own (which seems cleaner anyway).

BUG=342271

Review URL: https://codereview.chromium.org/228553005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@264119 4ff67af0-8c30-449e-8e8b-ad334ec8d88c