Refactors WebsiteSettingsBubbleController to use PermissionMenuModel.
In so doing, moves common code into a new base class, PermissionSelector,
which can be used by both views and cocoa UI code to define the menu for
each permission.

I'd like to take this farther and move the call to OnSitePermissionChanged() to
PermissionSelector.  It would require the PermissionSelector keeping a (weak)
pointer to the WebsiteSettings object - is that a reasonable change to make?

The two subclasses for PermissionSelector deviate a bit - for views, the
PermissionSelectorView contains all of the UI for displaying the whole line. 
For cocoa, it only acts as a bridge between the menu model and the UI for the
menu.

BUG=None

Review URL: https://codereview.chromium.org/193383002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@264329 4ff67af0-8c30-449e-8e8b-ad334ec8d88c