[fsp] Add an initial AsyncFileUtil.
This patch adds the AsyncFileUtil implementation for the file system provider.
It does not communicate with the file system provider service yet.

Currently, for not supported operations SECURITY error is returned. For
allowed, but not implemented operations, it is NOT_FOUND error.

Unit tests are quite trivial at this stage, but they will grow once more
operations are implemented.

TEST=unit_test: FileSystemProviderAsyncFileUtilTest*
BUG=248427

Review URL: https://codereview.chromium.org/236303017

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@264462 4ff67af0-8c30-449e-8e8b-ad334ec8d88c