[Files.app] Make getDriveEntryProperties() take multiple entries instead of single entry
Call of this private API has some overhead between C++ in the browser process and JS in a renderer process. This patch enables the method take a multiple entries so we can request properties of multiple entries with one call, and minimize a number of overheads. 

This patch should change no functionality.

BUG=345196
TEST=browser_test passes.

Review URL: https://codereview.chromium.org/189093016

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@264476 4ff67af0-8c30-449e-8e8b-ad334ec8d88c