Fix thumbnails on Drive, when they are not uploaded yet.
When copying to Drive, files are first stored locally, and uploaded in the
background. Therefore, a user click on the file, before it is uploaded.

In such situation, the thumbnail url should be empty, since there is no
thumbnail on the server side yet. However, the thumbnailUrl was being
generated anyway. Since resource id was empty, the url was faulty, resulting
in a 404 error, hence missing thumbnail.

This patch checks if the resource id is empty. If so, then doesn't fill out
the thumbnail url. Files app JS code is capable of fallbacking in such
situations.

TEST=Tested manually by copying a big image to Drive and quickly clicking
     on it.
BUG=351211

Review URL: https://codereview.chromium.org/240583005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@264718 4ff67af0-8c30-449e-8e8b-ad334ec8d88c