[fsp] Add FileSystemURLParser to the file system provider.
The FileSystemURLParser is a class which parses a cracked FileSystemURL to
associate it with the proper ProvidedFileSystem class handling this URL.

This will be necessary to route operations from AsyncFileUtil to a proper
providing extension.

Along the way, the Service class has been fixed. Now, the provided file systems
are gracefully unmounted in destructor. As a result, the external mount points
are cleared neatly.

TEST=unit_tests: FileSystemProviderMountPathUtilTest*
BUG=248427

Review URL: https://codereview.chromium.org/237583015

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@264780 4ff67af0-8c30-449e-8e8b-ad334ec8d88c