Extract GetNodeByID() method from BookmarkModel.
This moves GetNodeByID() method out of BookmarkModel and into a
standalone function in bookmark_utils.

The reason is that GetNodeByID() does not use any internal state of
BookmarkModel, so there is no real reason it needs to be there.

Scott suggested this and is fine with moving it into bookmark_utils.

BUG=359565
TEST=unit_tests, sync_integration_tests, chrome.
R=sky@chromium.org

Review URL: https://codereview.chromium.org/242823002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@265201 4ff67af0-8c30-449e-8e8b-ad334ec8d88c