Add SM_CXSIZEFRAME and SM_CXPADDEDBORDER metrics to get window border size.
Vista added the SM_CXPADDEDBORDER system metric which needs to be included
in the border size. The previous code happened to work with 32-bit Chrome
because Windows added the padding into SM_CXSIZEFRAME (and set the padding to 0)
for backwards compatibility, but with 64-bit that changed.
GetSystemMetrics(SM_CXPADDEDBORDER) should always return 0 on XP.

BUG=361720

Review URL: https://codereview.chromium.org/243173005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@265329 4ff67af0-8c30-449e-8e8b-ad334ec8d88c