Add initial FakeServer verification for Sync tests
This CL adds a method to FakeServer to export current data using
base::Value. This data is consumed by a new class, FakeServerVerifier,
that provides methods for verifying the data in integration tests.

The initial verification is rather simple. For example, only entity
names are compared and bookmark folders are not considered.

BUG=NONE

Review URL: https://codereview.chromium.org/242143002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@265511 4ff67af0-8c30-449e-8e8b-ad334ec8d88c