Calls FireRefreshTokenRevoked if an account is removed.
We cannot reply on periodic reconcile actions to remove accounts from cookies. The reason is that we actually *don't* do it at that stage for Android. Instead, we are going to keep a list of accounts last seen by Chrome and store that in a shared pref. We will come to the old list to the new list and call FireRefreshTokenRevoked accordingingly.

BUG=305086

Review URL: https://codereview.chromium.org/213823004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@266108 4ff67af0-8c30-449e-8e8b-ad334ec8d88c