Passwords PSL Matching on Gnome: overwrite realm and origin
On LoginDatabase using platforms, a password entry found by a PSL match gets the signon realm and the origin overwritten to match the query password form. This way, when the credential is saved, it is saved for the new site.

On Gnome, the overwriting did not happen. This CL introduces this to match the LoginDatabase behaviour.

Also, on Gnome, PSL matching was used not only to get logins, but also to check for existing logins when adding and updating passwords. The latter two use cases should require exact matching (that's also what happens in LoginDatabase). This CL fixes that, and only keeps PSL matching for GetLogins.

BUG=364745
R=gcasto@chromium.org

Review URL: https://codereview.chromium.org/245563005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@266177 4ff67af0-8c30-449e-8e8b-ad334ec8d88c