Refactor the interface for the CrxDownloader to take a callback.
To be consistent with the rest of the artifacts in the component updater,
the non-blocking calls for the CrxDownloader take a callback to be called
when the non-blocking call completes.

This change allows adding further callbacks besides the completion
callback, such as a callback to be called when download progress is made.

This change is a pure refactoring and it should be a NOP in terms of
existing functionality.

BUG=366388

Review URL: https://codereview.chromium.org/250383008

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@266304 4ff67af0-8c30-449e-8e8b-ad334ec8d88c