End session on crash if current device-local account no longer exists
When a device-local account is removed, its policy is deleted from disk
immediately. If a session using this account happens to be in progress,
the session is allowed to continue with policy served from an in-memory
cache. If Chrome crashes later in the session, the policy becomes
completely unavailable. Exit the session in that case, rather than
allowing it to continue without policy.

BUG=178290
TEST=Manual in VM

Review URL: https://codereview.chromium.org/252563007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@266568 4ff67af0-8c30-449e-8e8b-ad334ec8d88c