Move DefaultActivationClient to wm/core
Several tests are using AuraTestHelper and were relying on it to set up the DefaultActivationClient, so I had to have them instantiate the DAC directly after calling SetUp() on the helper. I also had to make the DAC delete itself after the root window it's bound to is destroyed.

R=sky@chromium.org
http://crbug.com/308710

Review URL: https://codereview.chromium.org/250113002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@266616 4ff67af0-8c30-449e-8e8b-ad334ec8d88c