Reland of r266437: [Files.app] Use getDriveEntryProperties() to retrieve metadata
This is a reland of r260688, which is reverted due to last-minute rebase mistake. 

Previously, we used FileEntry.getMetadata() to retrieve filesystem metadata and getDriveEntryProperties() to get Drive metadata, so we need 2 calls for 1 file on Drive. 

With this patch, getDriveEntryProperties() returns not only Drive metadata but also filesystem metadata. It's enough to call only getDriveEntryProperties() and we can reduce a number of calls by half. 

BUG=345196
TEST=browser_test passes. 
R=asargent@chromium.org, hashimoto@chromium.org, hirono@chromium.org 
TBR=asargent@chromium.org, hashimoto@chromium.org, hirono@chromium.org 
NOTRY=True
# NOTRYing for buildbots already passes (see. crrev.com/256023002)

Review URL: https://codereview.chromium.org/256023002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@266795 4ff67af0-8c30-449e-8e8b-ad334ec8d88c