Don't upload extension IDs in the cloud policy protocol.
The cloud policy protocol can also be used to fetch policy for extensions; this change makes an update to the protocol such that extension IDs are never uploaded to the server.

Note that this feature is still disabled by default behind the --enable-component-cloud-policy flag.

BUG=361156

Review URL: https://codereview.chromium.org/233423002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@266857 4ff67af0-8c30-449e-8e8b-ad334ec8d88c