Convert KeyboardDrivenEventRewriter to a ui::EventRewriter.
KeyboardDrivenEventRewriter was previously called at the start of
chromeos::EventRewriter, and suppressed further rewriting if it made
any changes. Since EventSource follows that model of breaking out of
the rewriter list, it makes sense to have this be a ui::EventRewriter
itself.

EventRewriterController will also be used for chromeos::EventRewriter
when that is converted to a ui::EventRewriter.

BUG=354034
TEST=unit_test
R=derat@chromium.org, sadrul@chromium.org

Review URL: https://codereview.chromium.org/240173007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@266969 4ff67af0-8c30-449e-8e8b-ad334ec8d88c