Password bubble: Create subviews for ManagePasswordsBubbleView.
Same idea as we've done for ManagePasswordItemView: separate subview
classes make it simpler to change the content and to understand where
the content is coming from.

After this CL, it will be a simpler addition to shoehorn in a blacklist
confirmation screen in the same way that we've done undo for the item
views.

BUG=261628

Review URL: https://codereview.chromium.org/255903006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@267213 4ff67af0-8c30-449e-8e8b-ad334ec8d88c