Refactor lots of stuff to use TemplateURLData instead of TemplateURL.
Only attach a Profile when it's needed.

This will allow methods for accessing the prepopulated search engines to be used before the Profile is available (i.e., as we are loading prefs). Which is required to determine if the current DSE in Prefs is user-selected or came from pre-populated.

BUG=365762
R=engedy@chromium.org, pkasting@chromium.org

Review URL: https://codereview.chromium.org/253573004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@267262 4ff67af0-8c30-449e-8e8b-ad334ec8d88c