Move Cocoa website_settings_bubble_controller* to website_settings/.
Adds website_settings_utils_cocoa* so that title-layout functionality
can be shared between PermissionBubbleController and PermissionSelectorButton.

Fixes bug where the title of a permission menu button in the permission
bubble might be cut off when the permission is changed.  Now, it re-calculates
the correct frame for the title when the permission changes.

This fix is included in this CL because it uses the newly-shared function in
website_settings_utils_cocoa.

BUG=None

Review URL: https://codereview.chromium.org/260813003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@267451 4ff67af0-8c30-449e-8e8b-ad334ec8d88c