cc: Add initial GPU-to-GPU copy rasterizer implementation.
This adds a derived RasterWorkerPool class that use a smaller
resource pool for staging and initialize tile resources by
copying the contents of a staging resource to a tile resource
using CHROMIUM_copy_texture.

Zero-copy is still used by default when map-image rasterizer
is enabled using --enable-map-image. GPU-to-GPU copy
rasterizer is used when --disable-zero-copy switch is
provided in combination with --enable-map-image. This
zero-copy setting is also added to about:flags.

BUG=269808
TBR=sky

Review URL: https://codereview.chromium.org/236313006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@267455 4ff67af0-8c30-449e-8e8b-ad334ec8d88c