[safe-browsing] Database full hash matches like prefix match.
The safe-browsing code incorrectly replicates full-hashes into the
prefix set, and treats database full hashes as a form of cached
response.  Move the database full hashes into PrefixSet in preparation
for removing the replication.  Treat full-hash matches and prefix
matches the same (as matches which may lead to a gethash request).
ContainsBrowseUrl() full hash result changed to indicate that it is
previously-cached results.

Add a unit test that a full hash is persisted by PrefixSet.  Right now
it succeeds because of the replication of full-hash prefixes into the
prefix set.

BUG=361248,172527

Review URL: https://codereview.chromium.org/257383006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@267498 4ff67af0-8c30-449e-8e8b-ad334ec8d88c