Fixed ordering assumption bug in AppInfoPermissionsTabTest
Fixed AppInfoPermissionsTabTest::RetainedFilePermissionsObtainedCorrectly test
to no longer assume that retained files are returned in a particular order.

Since the files are read as keys of a dictionary in
SavedFilesService::GetAllFileEntries(), there are no guarantees on the ordering
of the files, and this is now reflected in the test.

BUG=NONE

Review URL: https://codereview.chromium.org/261753003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@267509 4ff67af0-8c30-449e-8e8b-ad334ec8d88c