Use process_singleton_linux on Mac.
The current process singleton implementation on Linux does what we want for
Mac. I.e. it creates a unix domain socket in a temporary directory and a symlink
to it in the user data dir. The motivation for this in Linux is to support having
the user data dir on a networked filesystem. On Mac, it circumvents the path
length limit of unix domain sockets (104 characters). A similar solution is used
to allow app shims to connect to Chrome.

BUG=148465,353047,367612

Review URL: https://codereview.chromium.org/218883008

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@267512 4ff67af0-8c30-449e-8e8b-ad334ec8d88c