Download the high-res avatars when using the --new-profile-management flag
The high res avatar is downloaded:
- when we launch a new browser for a profile
- when we switch the avatar for a profile.

For now we re-download the file every time one of these actions is taken, 
regardless of whether the file exists on disk. Since this is still in development
and the avatar resources keep changing, we want to make sure someone isn't
stuck with an ugly resource. There's a todo in the code to stop doing this as
soon as the resources settle a bit :)

BUG=305048
TEST=Start Chrome with --new-profile-management. Create a new profile. 
In the avatar bubble, the active profile card should show a high-res (gray
 background) avatar. If you go to the user manager, the avatar displayed
 should be the same high res one (and not at all pixelated one). To double
 check, the file for that avatar should have been added to
 $USER_DATA_DIR/Avatars/

Review URL: https://codereview.chromium.org/222313005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@267529 4ff67af0-8c30-449e-8e8b-ad334ec8d88c